+++ catalog/app/components/FileEditor/index.ts
@@ -1,1 +1,2 @@
 export * from './FileEditor'
+export { default as Controls } from './Controls'
