+++ catalog/app/components/FileEditor/types.ts
@@ -1,0 +1,5 @@
+export type Mode = 'markdown' | 'text' | 'yaml'
+
+export interface EditorInputType {
+  brace: Mode | null
+}
